+++ components/form-item/index.ts
@@ -1,5 +1,5 @@
 import GFormItem from "./FormItem.vue";
-import { useFormItem } from '@flash-global66/b2b-ui-form';
+import { useFormItem } from '@flash-global66/g-form';
 
 type GFormItemInstance = InstanceType<typeof GFormItem>