+++ components/select/src/hooks/use-select.ts
@@ -18,10 +18,10 @@ import {
   useNamespace
 } from 'element-plus'
 import { CHANGE_EVENT, EVENT_CODE, UPDATE_MODEL_EVENT } from 'element-plus/es/constants/index.mjs'
-import { useFormItem, useFormItemInputId } from '@flash-global66/g-form/index.js';
+import { useFormItem, useFormItemInputId } from '@flash-global66/g-form'
 
-import { useAllowCreate } from './useAllowCreate'
-import { useProps } from './useProps'
+import { useAllowCreate } from './use-allow-create'
+import { useProps } from './use-props'
 
 import type { Option, OptionType, SelectStates } from '../types/select.types'
 import type { ISelectProps } from '../types/token'