+++ components/tag/index.ts
@@ -1,10 +1,10 @@
 import { withInstall, SFCWithInstall } from 'element-plus/es/utils/index'
 
-import Tag from './tag.vue'
+import Tag from './src/tag.vue'
 
 export const GTag: SFCWithInstall<typeof Tag> = withInstall(Tag)
 export type TagInstance = InstanceType<typeof Tag>
 
-export * from './tag'
+export * from './src/tag'
 
 export default GTag