+++ components/checkbox/CHANGELOG.md
@@ -3,6 +3,52 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.1.0](https://github.com/Flash-Global66/global-design-system/compare/@flash-global66/g-checkbox@0.0.11...@flash-global66/g-checkbox@0.1.0) (2025-04-01)
+
+
+### Features
+
+* **deps:** Add peer dependencies to components ([69ca8a6](https://github.com/Flash-Global66/global-design-system/commit/69ca8a6c26e2fd2777d5a6dea7cc9e7c5a0f9616))
+
+
+
+
+
+## [0.0.11](https://github.com/Flash-Global66/global-design-system/compare/@flash-global66/g-checkbox@0.0.10...@flash-global66/g-checkbox@0.0.11) (2025-04-01)
+
+
+### Bug Fixes
+
+* **checkbox:** Update checkbox styles to use element modifier syntax ([1552bf6](https://github.com/Flash-Global66/global-design-system/commit/1552bf6d84ca08545d237d8eaaa0d11b50139a72))
+
+
+
+
+
+## [0.0.10](https://github.com/Flash-Global66/global-design-system/compare/@flash-global66/g-checkbox@0.0.9...@flash-global66/g-checkbox@0.0.10) (2025-04-01)
+
+**Note:** Version bump only for package @flash-global66/g-checkbox
+
+
+
+
+
+## [0.0.9](https://github.com/Flash-Global66/global-design-system/compare/@flash-global66/g-checkbox@0.0.8...@flash-global66/g-checkbox@0.0.9) (2025-03-31)
+
+**Note:** Version bump only for package @flash-global66/g-checkbox
+
+
+
+
+
+## [0.0.8](https://github.com/Flash-Global66/global-design-system/compare/@flash-global66/g-checkbox@0.0.7...@flash-global66/g-checkbox@0.0.8) (2025-03-31)
+
+**Note:** Version bump only for package @flash-global66/g-checkbox
+
+
+
+
+
 ## 0.0.7 (2025-03-28)