+++ components/input/CHANGELOG.md
@@ -3,6 +3,47 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.3.0](https://github.com/Flash-Global66/global-design-system/compare/@flash-global66/g-input@0.2.0...@flash-global66/g-input@0.3.0) (2025-04-01)
+
+
+### Features
+
+* **deps:** Add peer dependencies to components ([69ca8a6](https://github.com/Flash-Global66/global-design-system/commit/69ca8a6c26e2fd2777d5a6dea7cc9e7c5a0f9616))
+
+
+
+
+
+# [0.2.0](https://github.com/Flash-Global66/global-design-system/compare/@flash-global66/g-input@0.1.1...@flash-global66/g-input@0.2.0) (2025-03-31)
+
+
+### Features
+
+* **form:** rename B2BForm to GForm and update build config ([c3a33b1](https://github.com/Flash-Global66/global-design-system/commit/c3a33b179ebdcae0dd9a9f174d27ab491366a16d))
+
+
+
+
+
+## [0.1.1](https://github.com/Flash-Global66/global-design-system/compare/@flash-global66/g-input@0.1.0...@flash-global66/g-input@0.1.1) (2025-03-31)
+
+**Note:** Version bump only for package @flash-global66/g-input
+
+
+
+
+
+# [0.1.0](https://github.com/Flash-Global66/global-design-system/compare/@flash-global66/g-input@0.0.7...@flash-global66/g-input@0.1.0) (2025-03-31)
+
+
+### Features
+
+* **input:** refactor input component ([64385d6](https://github.com/Flash-Global66/global-design-system/commit/64385d6381878c139e308c7d86538b84679b45cf))
+
+
+
+
+
 ## 0.0.7 (2025-03-28)