+++ components/radio/CHANGELOG.md
@@ -3,6 +3,47 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.2.0](https://github.com/Flash-Global66/global-design-system/compare/@flash-global66/g-radio@0.1.1...@flash-global66/g-radio@0.2.0) (2025-04-01)
+
+
+### Features
+
+* **deps:** Add peer dependencies to components ([69ca8a6](https://github.com/Flash-Global66/global-design-system/commit/69ca8a6c26e2fd2777d5a6dea7cc9e7c5a0f9616))
+
+
+
+
+
+## [0.1.1](https://github.com/Flash-Global66/global-design-system/compare/@flash-global66/g-radio@0.1.0...@flash-global66/g-radio@0.1.1) (2025-04-01)
+
+
+### Bug Fixes
+
+* **radio:** Update radio styles to use element modifier syntax ([36f6490](https://github.com/Flash-Global66/global-design-system/commit/36f6490d49fd6046f3b41c3a825ce136c1cfb9ab))
+
+
+
+
+
+# [0.1.0](https://github.com/Flash-Global66/global-design-system/compare/@flash-global66/g-radio@0.0.8...@flash-global66/g-radio@0.1.0) (2025-03-31)
+
+
+### Features
+
+* **radio:** Update radio component tsconfig ([5f56e27](https://github.com/Flash-Global66/global-design-system/commit/5f56e276a0c19a1ac8ef02c4cc24e232c292f549))
+
+
+
+
+
+## [0.0.8](https://github.com/Flash-Global66/global-design-system/compare/@flash-global66/g-radio@0.0.7...@flash-global66/g-radio@0.0.8) (2025-03-31)
+
+**Note:** Version bump only for package @flash-global66/g-radio
+
+
+
+
+
 ## 0.0.7 (2025-03-28)