+++ components/segmented/CHANGELOG.md
@@ -3,6 +3,47 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.2.0](https://github.com/Flash-Global66/global-design-system/compare/@flash-global66/g-segmented@0.1.1...@flash-global66/g-segmented@0.2.0) (2025-04-01)
+
+
+### Features
+
+* **deps:** Add peer dependencies to components ([69ca8a6](https://github.com/Flash-Global66/global-design-system/commit/69ca8a6c26e2fd2777d5a6dea7cc9e7c5a0f9616))
+
+
+
+
+
+## [0.1.1](https://github.com/Flash-Global66/global-design-system/compare/@flash-global66/g-segmented@0.1.0...@flash-global66/g-segmented@0.1.1) (2025-04-01)
+
+
+### Bug Fixes
+
+* **segmented:** Fix help-error style ([487ee89](https://github.com/Flash-Global66/global-design-system/commit/487ee890eeacf8bc2b05189ef6941a43d52669ff))
+
+
+
+
+
+# [0.1.0](https://github.com/Flash-Global66/global-design-system/compare/@flash-global66/g-segmented@0.0.9...@flash-global66/g-segmented@0.1.0) (2025-03-31)
+
+
+### Features
+
+* **segmented:** Add help text and error handling ([0e6658e](https://github.com/Flash-Global66/global-design-system/commit/0e6658e8d14e4bed0b6cdb48b42fca942192c84d))
+
+
+
+
+
+## [0.0.9](https://github.com/Flash-Global66/global-design-system/compare/@flash-global66/g-segmented@0.0.8...@flash-global66/g-segmented@0.0.9) (2025-03-31)
+
+**Note:** Version bump only for package @flash-global66/g-segmented
+
+
+
+
+
 ## 0.0.8 (2025-03-28)