+++ components/select/CHANGELOG.md
@@ -3,6 +3,66 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.1](https://github.com/Flash-Global66/global-design-system/compare/@flash-global66/g-select@0.1.0...@flash-global66/g-select@0.1.1) (2025-04-02)
+
+
+### Bug Fixes
+
+* improvement to styles in tag ([b7d6273](https://github.com/Flash-Global66/global-design-system/commit/b7d6273e7311eb96925870a2622396f5661dc930))
+
+
+
+
+
+# [0.1.0](https://github.com/Flash-Global66/global-design-system/compare/@flash-global66/g-select@0.0.11...@flash-global66/g-select@0.1.0) (2025-04-01)
+
+
+### Features
+
+* **deps:** Add peer dependencies to components ([69ca8a6](https://github.com/Flash-Global66/global-design-system/commit/69ca8a6c26e2fd2777d5a6dea7cc9e7c5a0f9616))
+
+
+
+
+
+## [0.0.11](https://github.com/Flash-Global66/global-design-system/compare/@flash-global66/g-select@0.0.10...@flash-global66/g-select@0.0.11) (2025-04-01)
+
+
+### Bug Fixes
+
+* **select:** Update select styles to use element modifier syntax ([e657fca](https://github.com/Flash-Global66/global-design-system/commit/e657fca9d4f53d17df780d8c9583bec8fb0591af))
+
+
+
+
+
+## [0.0.10](https://github.com/Flash-Global66/global-design-system/compare/@flash-global66/g-select@0.0.9...@flash-global66/g-select@0.0.10) (2025-04-01)
+
+
+### Bug Fixes
+
+* **select:** Fix empty state style selector ([112e961](https://github.com/Flash-Global66/global-design-system/commit/112e961536a96c1ac830439e475dcc15e7b406ba))
+
+
+
+
+
+## [0.0.9](https://github.com/Flash-Global66/global-design-system/compare/@flash-global66/g-select@0.0.8...@flash-global66/g-select@0.0.9) (2025-03-31)
+
+**Note:** Version bump only for package @flash-global66/g-select
+
+
+
+
+
+## [0.0.8](https://github.com/Flash-Global66/global-design-system/compare/@flash-global66/g-select@0.0.7...@flash-global66/g-select@0.0.8) (2025-03-31)
+
+**Note:** Version bump only for package @flash-global66/g-select
+
+
+
+
+
 ## 0.0.7 (2025-03-28)