+++ components/tag/CHANGELOG.md
@@ -3,6 +3,66 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.1](https://github.com/Flash-Global66/global-design-system/compare/@flash-global66/g-tag@0.2.0...@flash-global66/g-tag@0.2.1) (2025-04-02)
+
+
+### Bug Fixes
+
+* improvement to styles in tag ([b7d6273](https://github.com/Flash-Global66/global-design-system/commit/b7d6273e7311eb96925870a2622396f5661dc930))
+
+
+
+
+
+# [0.2.0](https://github.com/Flash-Global66/global-design-system/compare/@flash-global66/g-tag@0.1.2...@flash-global66/g-tag@0.2.0) (2025-04-01)
+
+
+### Features
+
+* **deps:** Add peer dependencies to components ([69ca8a6](https://github.com/Flash-Global66/global-design-system/commit/69ca8a6c26e2fd2777d5a6dea7cc9e7c5a0f9616))
+
+
+
+
+
+## [0.1.2](https://github.com/Flash-Global66/global-design-system/compare/@flash-global66/g-tag@0.1.1...@flash-global66/g-tag@0.1.2) (2025-04-01)
+
+
+### Bug Fixes
+
+* **tag:** Update tag styles to use element modifier syntax ([b9c7c1d](https://github.com/Flash-Global66/global-design-system/commit/b9c7c1de97bb5757cbb8c6a933972b5f678ae094))
+
+
+
+
+
+## [0.1.1](https://github.com/Flash-Global66/global-design-system/compare/@flash-global66/g-tag@0.1.0...@flash-global66/g-tag@0.1.1) (2025-04-01)
+
+**Note:** Version bump only for package @flash-global66/g-tag
+
+
+
+
+
+# [0.1.0](https://github.com/Flash-Global66/global-design-system/compare/@flash-global66/g-tag@0.0.8...@flash-global66/g-tag@0.1.0) (2025-03-31)
+
+
+### Features
+
+* **tag:** Implement tag component ([7b15a63](https://github.com/Flash-Global66/global-design-system/commit/7b15a63d1c5630ad4174506819e2dee96266aaf0))
+
+
+
+
+
+## [0.0.8](https://github.com/Flash-Global66/global-design-system/compare/@flash-global66/g-tag@0.0.7...@flash-global66/g-tag@0.0.8) (2025-03-31)
+
+**Note:** Version bump only for package @flash-global66/g-tag
+
+
+
+
+
 ## 0.0.7 (2025-03-28)